Split 'cardmodel' into a semi-MVC structure
The "view" is now in webclient/views/studyview.  It contains no logic at the
moment, only type definitions for view elements (buttons mostly).

The model is now the `repository` directory.  The "controllers" are
per-model type, and live in controllers/*/, with their glue also in
`repository`.  This isn't ideal, but it seemed like a step in the right
direction.  I hope to find some better place to put scheduling logic at some
point.  But for now, it's in `repository/modelcontroller.go` along with a
FIXME note.